Extract LedgerToJson.cpp from LedgerToJson.h.
